Magnetometer power down mode ok.
Even the LSM303D not providing a power down mode to magnetometer,
this can be done with the T_ONLY mode, where the termometer still on
but the magnetometer is turned off. The disvantage is that, to keep
it similar to accelerometer usage, when setting a data rate to
magnetometer, it is necessary to check if magnetometer is on. In
accelerometer it is not necessary because the same field that turns
accelerometer down, also sets the data rate. So, if a data rate is
set, it leaves the power down mode.